+++ Maneuver/OperationEllipticize.cs
@@ -7,7 +7,7 @@ namespace KRPC.MechJeb.Maneuver {
 		private readonly object newPeA;
 
 		public OperationEllipticize() : base("OperationEllipticize") {
-			this.newApA = this.type.GetField("newPeA").GetValue(this.instance);
+			this.newApA = this.type.GetField("newApA").GetValue(this.instance);
 			this.newPeA = this.type.GetField("newPeA").GetValue(this.instance);
 		}