+++ docs/conf.py
@@ -3,7 +3,7 @@ project = "kRPC.MechJeb"
 copyright = "2018-2019, Genhis"
 author = "Genhis"
 
-version = "0.4.0"
+version = "0.5.0"
 release = version
 
 master_doc = "index"