+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 - ValueError when setting enum properties
 ([#4](https://github.com/Genhis/KRPC.MechJeb/pull/4))
+- OperationEllipticize apoapsis field pointed to periapsis instead
 
 ## [0.4.1] - 2019-05-11
 ### Fixed