+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.0] - 2019-06-11
 ### Added
 - **SmartASS** window
 - **Translatron** window ([#5](https://github.com/Genhis/KRPC.MechJeb/pull/5))
@@ -87,7 +89,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - **StagingController**
 - **TargetController** to get information about the current target
 
-[Unreleased]: https://github.com/Genhis/KRPC.MechJeb/compare/v0.4.1...HEAD
+[Unreleased]: https://github.com/Genhis/KRPC.MechJeb/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/Genhis/KRPC.MechJeb/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/Genhis/KRPC.MechJeb/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/Genhis/KRPC.MechJeb/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/Genhis/KRPC.MechJeb/compare/v0.2.0...v0.3.0